Reorganize some of the utils and add a once function
