+++ packages/utils/src/utils.ts
@@ -1,5 +1,5 @@
-import get from 'lodash.get';
 import kindOf from 'kind-of';
+import { Nil } from './interfaces';
 import {
     isString,
     toString,
@@ -7,11 +7,15 @@ import {
     trimAndToLower
 } from './strings';
 
-export const isNil = (input: any) => input == null;
-export const isNotNil = (input: any) => input != null;
+export function isNil<T>(input: T|Nil): input is Nil {
+    return input == null;
+}
+export function isNotNil<T>(input: T|Nil): boolean {
+    return input != null;
+}
 
 /** Check if an input is empty, similar to lodash.isEmpty */
-export function isEmpty<T>(val?: T): val is undefined {
+export function isEmpty<T>(val?: T|null|undefined): val is undefined {
     const _val = val as any;
     if (!_val) return true;
     if (typeof _val.size === 'number') return !_val.size;
@@ -132,83 +136,4 @@ export function parseList(input: any): string[] {
     }
 
     return strings.map((s) => s.trim()).filter((s) => !!s);
-}
-
-export function noop(..._args: any[]): any {}
-
-/**
- * A typesafe get function (will always return the correct type)
- *
- * **IMPORTANT** This does not behave like lodash.get,
- * it does not deal with dot notation (nested fields)
- * and it will use the default when dealing with OR statements
- */
-export function getField<V>(
-    input: undefined,
-    field: string,
-    defaultVal?: V
-): V;
-export function getField<T extends {}, P extends keyof T>(
-    input: T,
-    field: P
-): T[P];
-export function getField<T extends {}, P extends keyof T>(
-    input: T | undefined,
-    field: P
-): T[P];
-export function getField<T extends {}, P extends keyof T>(
-    input: T | undefined,
-    field: P,
-    defaultVal: never[]
-): T[P];
-export function getField<T extends {}, P extends keyof T, V>(
-    input: T | undefined,
-    field: P,
-    defaultVal: V
-): T[P] | V;
-export function getField<T extends {}, P extends keyof T, V extends T[P]>(
-    input: T | undefined,
-    field: P, defaultVal: V
-): T[P];
-export function getField<T, P extends keyof T, V>(
-    input: T,
-    field: P,
-    defaultVal?: V
-): any {
-    const result = get(input, field);
-    if (isBooleanLike(defaultVal)) {
-        if (result == null) return defaultVal;
-        return result;
-    }
-    return result || defaultVal;
-}
-
-function _getArgCacheKey(args: any[]): string {
-    const fixed = args.filter((a, i, arr) => {
-        if (a === undefined && arr.length === (i + 1)) return false;
-        return true;
-    });
-    try {
-        return JSON.stringify(fixed);
-    } catch (_e) {
-        return toString(fixed);
-    }
-}
-
-type MemoizeFn = (...args: any[]) => any;
-/**
- * A replacement for lodash memoize
-*/
-export function memoize<T extends MemoizeFn>(fn: T): T {
-    const _cache = new Map<string, any>();
-
-    const _memoize: any = (...args: any[]): any => {
-        const key = _getArgCacheKey(args);
-        const cached = _cache.get(key);
-        if (cached !== undefined) return cached;
-        const result = fn(...args);
-        _cache.set(key, result);
-        return result;
-    };
-    return _memoize;
 }
